Moved KNNAnomalyClassifierRegion.py -> knn_anomaly_classifier_region.py
